requirements.txt
ollama
python-docx
chromadb
rank_bm25
